Подправил mapper, поднял версию базы данных, также удалил TimeUtils теперь преобразование из Long в String происходит в mapper
