+++ app/src/main/java/com/example/cryptoapp/data/database/model/CoinInfoDbModel.kt
@@ -12,5 +12,5 @@ data class CoinInfoDbModel(
     val highDay: String?,
     val lowDay: String?,
     val lastMarket: String?,
-    val imageUrl: String?
+    val imageUrl: String
 )
