+++ app/src/main/java/com/example/cryptoapp/presentation/view/CoinDetailActivity.kt
@@ -2,15 +2,12 @@ package com.example.cryptoapp.presentation.view
 
 import android.content.Context
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
-import androidx.lifecycle.Observer
+import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
-import com.example.cryptoapp.data.network.ApiFactory.BASE_IMAGE_URl
 import com.example.cryptoapp.databinding.ActivityCoinDetailBinding
 import com.example.cryptoapp.presentation.viewmodel.CoinViewModel
-import com.example.cryptoapp.utils.convertTimestampToTime
 import com.squareup.picasso.Picasso
 
 class CoinDetailActivity : AppCompatActivity() {
@@ -28,25 +25,26 @@ class CoinDetailActivity : AppCompatActivity() {
             finish()
             return
         }
-        val fromSymbol = intent.getStringExtra(EXTRA_FROM_SYMBOL)
+        val fromSymbol = intent.getStringExtra(EXTRA_FROM_SYMBOL) ?: EMPTY_SYMBOL
         viewModel = ViewModelProvider(this)[CoinViewModel::class.java]
-        viewModel.getDetailInfo(fromSymbol?: "").observe(this) {
+        viewModel.getDetailInfo(fromSymbol).observe(this) {
             Log.d("Detail_Info", it.toString())
             with(binding) {
                 tvPrice.text = it.price
                 tvMinPrice.text = it.lowDay
                 tvMaxPrice.text = it.highDay
                 tvLastMarket.text = it.lastMarket
-                tvUpdate.text = convertTimestampToTime(it.lastUpdate)
+                tvUpdate.text = it.lastUpdate
                 tvCoin.text = it.fromSymbol
                 tvToSymbol.text = it.toSymbol
-                Picasso.get().load(BASE_IMAGE_URl + it.imageUrl).into(ivCoinLogo)
+                Picasso.get().load(it.imageUrl).into(ivCoinLogo)
             }
         }
     }
 
     companion object {
         private const val EXTRA_FROM_SYMBOL = "fSym"
+        private const val EMPTY_SYMBOL = ""
 
         fun newIntent(context: Context, fromSymbol: String): Intent {
             val intent = Intent(context, CoinDetailActivity::class.java)
