--- app/src/main/java/com/example/cryptoapp/utils/TimeUtils.kt
@@ -1,15 +1,0 @@
-package com.example.cryptoapp.utils
-
-import java.sql.Timestamp
-import java.text.SimpleDateFormat
-import java.util.*
-
-fun convertTimestampToTime(timestamp: Long?): String {
-    if (timestamp == null) return ""
-    val stamp = Timestamp(timestamp * 1000)
-    val date = Date(stamp.time)
-    val pattern = "HH:mm:ss"
-    val simpleDateFormat = SimpleDateFormat(pattern, Locale.getDefault())
-    simpleDateFormat.timeZone = TimeZone.getDefault()
-    return simpleDateFormat.format(date)
-}
